backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-groq==0.3.8
langchain-huggingface==0.3.1
faiss-cpu==1.12.0
pypdf==6.0.0
sentence-transformers==5.1.0
